+++ Runtime/Scripts/ExpectedTypesOfResponse/FileResponse.cs
@@ -1,4 +1,6 @@
-﻿namespace KindMen.Uxios.ExpectedTypesOfResponse
+﻿using KindMen.Uxios.Http;
+
+namespace KindMen.Uxios.ExpectedTypesOfResponse
 {
     public sealed class FileResponse : ExpectedTypeOfResponse
     {
@@ -6,7 +8,7 @@
 
         public override void AddMetadataToRequest(Request request)
         {
-            request.Headers.TryAdd("Accept", "application/octet-stream");
+            request.Headers.TryAdd(Headers.Accept, "application/octet-stream");
         }
     }
 }