+++ Runtime/Scripts/ExpectedTypesOfResponse/TextResponse.cs
@@ -1,4 +1,4 @@
-﻿using System;
+﻿using KindMen.Uxios.Http;
 
 namespace KindMen.Uxios.ExpectedTypesOfResponse
 {
@@ -6,7 +6,7 @@ namespace KindMen.Uxios.ExpectedTypesOfResponse
     {
         public override void AddMetadataToRequest(Request request)
         {
-            request.Headers.TryAdd("Accept", "text/*");
+            request.Headers.TryAdd(Headers.Accept, "text/*");
         }
     }
 }