+++ Runtime/Scripts/ExpectedTypesOfResponse/TextureResponse.cs
@@ -1,4 +1,6 @@
-﻿namespace KindMen.Uxios.ExpectedTypesOfResponse
+﻿using KindMen.Uxios.Http;
+
+namespace KindMen.Uxios.ExpectedTypesOfResponse
 {
     public sealed class TextureResponse : ExpectedTypeOfResponse
     {
@@ -6,7 +8,7 @@
 
         public override void AddMetadataToRequest(Request request)
         {
-            request.Headers.TryAdd("Accept", "image/*");
+            request.Headers.TryAdd(Headers.Accept, "image/*");
         }
     }
 }